Run hooks on startup if the IP address is fresh
If the address was acquired within 30 seconds of starting up the ipwatch
program, still run the configured hooks. This helps in the situation
where a host may be starting ipwatch and some dynamic network
configuration within a close time period (e.g. on startup or a fresh
deployment that requires restarting both ipwatch and a DHCP client).